+++ database/migrations/2025_06_06_071130_add_date_transactionin_cast_account.php
@@ -8,9 +8,11 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('account_transactions', function (Blueprint $table) {
-            $table->date('date_transaction')->after('cast_account_destination_id');
-        });
+        if (!Schema::hasColumn('account_transactions', 'date_transaction')) {
+            Schema::table('account_transactions', function (Blueprint $table) {
+                $table->date('date_transaction')->after('cast_account_destination_id');
+            });
+        }
     }
 
     /**
@@ -18,8 +20,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('account_transactions', function (Blueprint $table) {
-            $table->dropColumn('date_transaction');
-        });
+        if (Schema::hasColumn('account_transactions', 'date_transaction')){
+            Schema::table('account_transactions', function (Blueprint $table) {
+                $table->dropColumn('date_transaction');
+            });
+        }
     }
 };
